post new signup complete. bug fix on navbar, sidebar, and profile.

+++ db/migrations/schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE users (
   name VARCHAR(255),
   password VARCHAR(255),
   email VARCHAR(255),
-  phone INTEGER,
+  phone BIGINT,
   username VARCHAR(255),
   date_created DATE DEFAULT CURRENT_DATE,
   xp INTEGER DEFAULT 0,
